grapher
really bad grapher that i made with the help from "curran kelleher"s tutorial "building a graphing calculator with HTML5". its slightly better that whats in there. also, it has minimal support for variables (they can only go from 0 to 100 for now, because thats the default range for an input tag range). as for the equations, i just mathjs to parse them, since i could not write a good parser for that, and much less implement it cleanly into this abomination.