slightly updated version of new grapher.
i decided to break the js code into three files. equation_handler deals with most of the user interactions with the interface, for example, when they type something into one of the equations, equation_handler will get the innerhtml and will draw_curve. graph_new is going to have the functions needed to draw to the plot. parser is going to be a parser i'm going to write, that way i can customize it, and it wont depend on other things in case they become deprecated.