I FORGOT TO CHECK IF IT WORKED
there was a piece of code i wrote at the top that was just for testing and i forgot to delete it. also one of the "this" were a "touchlist" for some reason.
also i forgot to mention that the parser still doesn't work. you can only replace the expression for now, but looking at the ast tree is pretty cool.